split beacon front/admin into separate blueprints
