--- purchasing/opportunities/views/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from purchasing.opportunities.views.admin import new, edit
-from purchasing.opportunities.views.front import (
-    index, signup, manage, browse, detail
-)
