+++ purchasing/templates/includes/nav.html
@@ -42,7 +42,7 @@
                 </span>
                 Sherpa
               </a></li>
-              <li><a href="{{ url_for('opportunities.index') }}">
+              <li><a href="{{ url_for('opportunities.splash') }}">
                 <span class="fa-stack fa-1x">
                   <i class="fa fa-envelope"></i>
                 </span>
