+++ purchasing/templates/opportunities/admin/opportunity.html
@@ -5,7 +5,7 @@
 <div class="container">
   <div class="row">
     <div class="col-md-8 col-md-offset-2">
-      <form method="POST" enctype="multipart/form-data" action="{% if opportunity.id %}{{ url_for('opportunities.edit', opportunity_id=opportunity.id) }}{% else %}{{ url_for('opportunities.new') }}{% endif %}">
+      <form method="POST" enctype="multipart/form-data" action="{% if opportunity.id %}{{ url_for('opportunities_admin.edit', opportunity_id=opportunity.id) }}{% else %}{{ url_for('opportunities_admin.new') }}{% endif %}">
 
         {{ form.csrf_token() }}
         {% include "includes/flashes.html" %}
@@ -110,21 +110,7 @@
         </div>
         {% endif %}
 
-
-        <div class="row">
-          <div class="col-sm-6">
-            <button type="submit" class="btn btn-primary col-sm-12" id="button-save">Save Opportunity</button>
-          </div>
-          <div class="col-sm-6">
-            <button type="submit" class="btn btn-primary col-sm-12" id="button-submit">
-              {% if current_user.role.name in ['conductor', 'admin', 'superadmin'] %}
-              Publish Opportunity
-              {% else %}
-              Submit Opportunity for Review
-              {% endif %}
-            </button>
-          </div>
-        </div>
+        <button type="submit" class="btn btn-primary col-sm-6">Publish Opportunity</button>
 
       </form>
