+++ purchasing/templates/opportunities/browse.html
@@ -49,7 +49,7 @@
               <td>{{ opportunity.department }}</td>
               <td>{{ opportunity.planned_deadline.strftime('%Y-%m-%d') }}</td>
               {% if opportunity.can_edit(current_user) %}
-              <td><a href="{{ url_for('opportunities.edit', opportunity_id=opportunity.id) }}">Edit</a></td>
+              <td><a href="{{ url_for('opportunities_admin.edit', opportunity_id=opportunity.id) }}">Edit</a></td>
               {% elif not current_user.is_anonymous() %}
               <td><i class="fa fa-minus"></i></td>
               {% endif %}
@@ -96,7 +96,7 @@
               <td>{{ opportunity.department }}</td>
               <td>{{ opportunity.planned_deadline.strftime('%Y-%m-%d') }}</td>
               {% if opportunity.can_edit(current_user) %}
-              <td><a href="{{ url_for('opportunities.edit', opportunity_id=opportunity.id) }}">Edit</a></td>
+              <td><a href="{{ url_for('opportunities_admin.edit', opportunity_id=opportunity.id) }}">Edit</a></td>
               {% elif not current_user.is_anonymous() %}
               <i class="fa fa-minus"></i>
               {% endif %}
