+++ purchasing/templates/opportunities/front/manage.html
