+++ purchasing/templates/opportunities/front/signup.html
