+++ purchasing/templates/opportunities/front/splash.html
