+++ purchasing/templates/opportunities/nav.html
@@ -1,6 +1,6 @@
 {% extends "includes/nav.html" %}
 
-{% block brandhref %}{{ url_for('opportunities.index') }}{% endblock %}
+{% block brandhref %}{{ url_for('opportunities.splash') }}{% endblock %}
 
 {% block brand %}Beacon{% endblock %}
 
@@ -9,7 +9,7 @@
 
 {% block applinks_right %}
 {% if not current_user.is_anonymous() %}
-<li><p class="navbar-text"><a href="{{ url_for('opportunities.new') }}" class="navbar-link">Add Opportunity</a></p></li>
+<li><p class="navbar-text"><a href="{{ url_for('opportunities_admin.new') }}" class="navbar-link">Add Opportunity</a></p></li>
 {% endif %}
 <li><p class="navbar-text"><a href="{{ url_for('opportunities.browse') }}" class="navbar-link">Browse Opportunities</a></p></li>
 <li><p class="navbar-text"><a href="{{ url_for('opportunities.signup') }}" class="navbar-link">Sign Up</a></p></li>
